Target position latency compensation fixed

+++ src/modules/mc_pos_control/mc_pos_control_main.cpp
@@ -759,7 +759,7 @@ MulticopterPositionControl::task_main()
 					}
 
 					/* calculate position setpoint with position prediction using GPS time */
-					float target_dt = math::constrain((_global_pos.time_gps_usec - _target_pos.time_gps_usec) / 1000000.0f, 0.0f, 1.0f);
+					float target_dt = math::constrain(((int64_t)_global_pos.time_gps_usec - (int64_t)_target_pos.time_gps_usec) / 1000000.0f, 0.0f, 1.0f);
 
 					add_vector_to_global_position(_target_pos.lat, _target_pos.lon,
 							_follow_offset(0) + _target_pos.vel_n * target_dt, _follow_offset(1) + _target_pos.vel_e * target_dt,
